# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/letterboxd/test": {
      "get": {
        "operationId": "LetterboxdController_testFeed",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "description": "Letterboxd username to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sample of parsed Letterboxd activities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterboxdTestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Username query parameter is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Test Letterboxd RSS feed for a user",
        "tags": [
          "letterboxd"
        ]
      }
    },
    "/letterboxd/poll": {
      "post": {
        "operationId": "LetterboxdController_triggerPoll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Polling started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterboxdPollingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually trigger RSS polling for all areas",
        "tags": [
          "letterboxd"
        ]
      }
    },
    "/github/test": {
      "get": {
        "operationId": "GithubController_testRepository",
        "parameters": [
          {
            "name": "owner",
            "required": true,
            "in": "query",
            "description": "Organisation ou utilisateur GitHub",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "required": true,
            "in": "query",
            "description": "Nom du dépôt",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Clé d’action optionnelle pour filtrer le résultat",
            "schema": {
              "enum": [
                "new_issue",
                "new_pull_request",
                "new_release"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Récupération manuelle effectuée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubTestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Paramètres manquants ou action non prise en charge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tester la récupération GitHub pour un dépôt",
        "tags": [
          "github"
        ]
      }
    },
    "/github/poll": {
      "post": {
        "operationId": "GithubController_triggerPolling",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Polling lancé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubPollingResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Déclencher manuellement le polling GitHub pour toutes les areas",
        "tags": [
          "github"
        ]
      }
    },
    "/discord/test/messages": {
      "get": {
        "description": "Retourne jusqu’à 50 messages récents pour le salon indiqué. Nécessite la variable DISCORD_BOT_TOKEN dans la configuration.",
        "operationId": "DiscordController_getRecentMessages",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "query",
            "description": "Identifiant du salon dont il faut récupérer les messages (obligatoire).",
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximum de messages à récupérer (1-50)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages récents du salon",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscordMessagePreviewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les derniers messages d’un salon",
        "tags": [
          "discord"
        ]
      }
    },
    "/discord/test/send-message": {
      "post": {
        "description": "Envoie un message texte simple via le bot configuré. Utile pour tester manuellement l’intégration Discord.",
        "operationId": "DiscordController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendDiscordMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message envoyé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envoyer un message dans un salon Discord",
        "tags": [
          "discord"
        ]
      }
    },
...
  },
  "info": {
    "title": "Area API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Action invalide",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "UnauthorizedResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Non autorisé",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Identifiants invalides",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "LetterboxdActivityPreviewDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "review",
              "diary",
              "watched",
              "list",
              "rating"
            ],
            "example": "review"
          },
          "filmTitle": {
            "type": "string",
            "example": "Interstellar"
          },
          "filmYear": {
            "type": "object",
            "example": 2014,
            "nullable": true
          },
          "rating": {
            "type": "object",
            "example": 4.5,
            "nullable": true
          },
          "reviewText": {
            "type": "object",
            "example": "Watched for the third time and it still blows my mind.",
            "nullable": true
          },
          "watchedDate": {
            "type": "object",
            "example": "2025-01-18T21:15:00.000Z",
            "nullable": true
          },
          "letterboxdUrl": {
            "type": "string",
            "example": "https://letterboxd.com/user/film/interstellar/"
          },
          "activityDate": {
            "type": "string",
            "example": "2025-01-19T09:30:00.000Z"
          },
          "isRewatch": {
            "type": "object",
            "example": true,
            "nullable": true
          }
        },
        "required": [
          "type",
          "filmTitle",
          "filmYear",
          "rating",
          "reviewText",
          "watchedDate",
          "letterboxdUrl",
          "activityDate",
          "isRewatch"
        ]
      },
      "LetterboxdTestResponseDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cinephile42"
          },
          "count": {
            "type": "number",
            "example": 25
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LetterboxdActivityPreviewDto"
            }
          }
        },
        "required": [
          "username",
          "count",
          "activities"
        ]
      },
      "LetterboxdPollingResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Polling triggered successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "GithubTestResponseDto": {
        "type": "object",
        "properties": {
          "owner": {
            "type": "string",
            "example": "open-source-org"
          },
          "repo": {
            "type": "string",
            "example": "area-backend"
          },
          "fetchedAt": {
            "type": "string",
            "example": "2025-01-20T12:34:56.000Z",
            "description": "Horodatage de l’exécution de la récupération manuelle"
          },
          "result": {
            "type": "object",
            "description": "Charge utile brute renvoyée par GitHub (spécifique à l’action)",
            "nullable": true
          }
        },
        "required": [
          "owner",
          "repo",
          "fetchedAt",
          "result"
        ]
      },
      "GithubPollingResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Polling GitHub lancé avec succès"
          }
        },
        "required": [
          "message"
        ]
      },
      "DiscordAuthorPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "example": "123456789012345678",
            "nullable": true
          },
          "username": {
            "type": "object",
            "example": "JaneDoe",
            "nullable": true
          },
          "globalName": {
            "type": "object",
            "example": "Jane Doe",
            "nullable": true
          },
          "displayName": {
            "type": "object",
            "example": "Jane",
            "nullable": true
          },
          "isBot": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "username",
          "globalName",
          "displayName",
          "isBot"
        ]
      },
      "DiscordAttachmentPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "113355779988776655"
          },
          "url": {
            "type": "string",
            "example": "https://cdn.discordapp.com/attachments/.../image.png"
          },
          "filename": {
            "type": "object",
            "example": "image.png",
            "nullable": true
          },
          "contentType": {
            "type": "object",
            "example": "image/png",
            "nullable": true
          }
        },
        "required": [
          "id",
          "url",
          "filename",
          "contentType"
        ]
      },
      "DiscordMessagePreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123456789012345678"
          },
          "channelId": {
            "type": "string",
            "example": "123456789012345678"
          },
          "guildId": {
            "type": "object",
            "example": "123456789012345678",
            "nullable": true
          },
          "content": {
            "type": "string",
            "example": "Ceci est le contenu d’un message d’exemple"
          },
          "author": {
            "$ref": "#/components/schemas/DiscordAuthorPreviewDto"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscordAttachmentPreviewDto"
            }
          },
          "mentions": {
            "example": [
              "123456789012345678"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2025-01-27T10:15:00.000Z"
          },
          "url": {
            "type": "string",
            "example": "https://discord.com/channels/123/456/789"
          }
        },
        "required": [
          "id",
          "channelId",
          "guildId",
          "content",
          "author",
          "attachments",
          "mentions",
          "createdAt",
          "url"
        ]
      },
      "SendDiscordMessageDto": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string",
            "example": "123456789012345678",
            "description": "Identifiant du salon cible (snowflake)"
          },
          "content": {
            "type": "string",
            "example": "Message envoyé automatiquement par Area !",
            "description": "Contenu texte brut du message (2000 caractères max.)",
            "maxLength": 2000
          },
          "mentionUserIds": {
            "description": "Liste optionnelle d’identifiants d’utilisateurs à mentionner",
            "example": [
              "123456789012345678"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mentionRoleIds": {
            "description": "Liste optionnelle d’identifiants de rôles à mentionner",
            "example": [
              "987654321987654321"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "channelId",
          "content"
        ]
      },
...
    }
  }
}